ref(processort): Split off replay processing code into separate sub-module
