+++ relay-server/src/actors/processor.rs
@@ -1,6 +1,5 @@
 use std::error::Error;
 use std::io::Write;
-use std::net::IpAddr as NetIPAddr;
 use std::ops::ControlFlow;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
@@ -16,8 +15,7 @@ use relay_auth::RelayVersion;
 use relay_base_schema::project::{ProjectId, ProjectKey};
 use relay_common::time::UnixTimestamp;
 use relay_config::{Config, HttpEncoding};
-use relay_dynamic_config::{ErrorBoundary, Feature, GlobalConfig, ProjectConfig};
-use relay_event_normalization::replay::{self, ReplayError};
+use relay_dynamic_config::{ErrorBoundary, Feature, GlobalConfig};
 use relay_event_normalization::{
     nel, ClockDriftProcessor, DynamicMeasurementsConfig, MeasurementsConfig,
     NormalizeProcessorConfig, TransactionNameConfig,
@@ -26,7 +24,7 @@ use relay_event_normalization::{GeoIpLookup, RawUserAgentInfo};
 use relay_event_schema::processor::{self, process_value, ProcessingAction, ProcessingState};
 use relay_event_schema::protocol::{
     Breadcrumb, Contexts, Csp, Event, EventType, ExpectCt, ExpectStaple, Hpkp, IpAddr,
-    LenientString, Metrics, NetworkReportError, OtelContext, ProfileContext, RelayInfo, Replay,
+    LenientString, Metrics, NetworkReportError, OtelContext, ProfileContext, RelayInfo,
     SecurityReportType, Timestamp, TraceContext, Values,
 };
 use relay_filter::FilterStatKey;
@@ -36,7 +34,6 @@ use relay_pii::{scrub_graphql, PiiAttachmentsProcessor, PiiConfigError, PiiProce
 use relay_profiling::{ProfileError, ProfileId};
 use relay_protocol::{Annotated, Array, Empty, FromValue, Object, Value};
 use relay_quotas::{DataCategory, Scoping};
-use relay_replays::recording::RecordingScrubber;
 use relay_sampling::config::{RuleType, SamplingMode};
 use relay_sampling::evaluation::{
     MatchedRuleIds, ReservoirCounters, ReservoirEvaluator, SamplingEvaluator,
@@ -74,6 +71,7 @@ use crate::utils::{
     self, ChunkedFormDataAggregator, FormDataIter, ItemAction, ManagedEnvelope, SamplingResult,
 };
 
+mod replay;
 mod report;
 mod session;
 
@@ -729,137 +727,6 @@ impl EnvelopeProcessorService {
                 }
             }
         }
-    }
-
-    /// Remove replays if the feature flag is not enabled.
-    fn process_replays(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
-        let project_state = &state.project_state;
-        let replays_enabled = project_state.has_feature(Feature::SessionReplay);
-        let scrubbing_enabled = project_state.has_feature(Feature::SessionReplayRecordingScrubbing);
-
-        let meta = state.envelope().meta().clone();
-        let client_addr = meta.client_addr();
-        let event_id = state.envelope().event_id();
-
-        let limit = self.inner.config.max_replay_uncompressed_size();
-        let config = project_state.config();
-        let datascrubbing_config = config
-            .datascrubbing_settings
-            .pii_config()
-            .map_err(|e| ProcessingError::PiiConfigError(e.clone()))?
-            .as_ref();
-        let mut scrubber =
-            RecordingScrubber::new(limit, config.pii_config.as_ref(), datascrubbing_config);
-
-        let user_agent = &RawUserAgentInfo {
-            user_agent: meta.user_agent(),
-            client_hints: meta.client_hints().as_deref(),
-        };
-
-        state.managed_envelope.retain_items(|item| match item.ty() {
-            ItemType::ReplayEvent => {
-                if !replays_enabled {
-                    return ItemAction::DropSilently;
-                }
-
-                match self.process_replay_event(&item.payload(), config, client_addr, user_agent) {
-                    Ok(replay) => match replay.to_json() {
-                        Ok(json) => {
-                            item.set_payload(ContentType::Json, json);
-                            ItemAction::Keep
-                        }
-                        Err(error) => {
-                            relay_log::error!(
-                                error = &error as &dyn Error,
-                                "failed to serialize replay"
-                            );
-                            ItemAction::Keep
-                        }
-                    },
-                    Err(error) => {
-                        relay_log::warn!(error = &error as &dyn Error, "invalid replay event");
-                        ItemAction::Drop(Outcome::Invalid(match error {
-                            ReplayError::NoContent => DiscardReason::InvalidReplayEventNoPayload,
-                            ReplayError::CouldNotScrub(_) => DiscardReason::InvalidReplayEventPii,
-                            ReplayError::CouldNotParse(_) => DiscardReason::InvalidReplayEvent,
-                            ReplayError::InvalidPayload(_) => DiscardReason::InvalidReplayEvent,
-                        }))
-                    }
-                }
-            }
-            ItemType::ReplayRecording => {
-                if !replays_enabled {
-                    return ItemAction::DropSilently;
-                }
-
-                // XXX: Processing is there just for data scrubbing. Skip the entire expensive
-                // processing step if we do not need to scrub.
-                if !scrubbing_enabled || scrubber.is_empty() {
-                    return ItemAction::Keep;
-                }
-
-                // Limit expansion of recordings to the max replay size. The payload is
-                // decompressed temporarily and then immediately re-compressed. However, to
-                // limit memory pressure, we use the replay limit as a good overall limit for
-                // allocations.
-                let parsed_recording = metric!(timer(RelayTimers::ReplayRecordingProcessing), {
-                    scrubber.process_recording(&item.payload())
-                });
-
-                match parsed_recording {
-                    Ok(recording) => {
-                        item.set_payload(ContentType::OctetStream, recording);
-                        ItemAction::Keep
-                    }
-                    Err(e) => {
-                        relay_log::warn!("replay-recording-event: {e} {event_id:?}");
-                        ItemAction::Drop(Outcome::Invalid(
-                            DiscardReason::InvalidReplayRecordingEvent,
-                        ))
-                    }
-                }
-            }
-            _ => ItemAction::Keep,
-        });
-
-        Ok(())
-    }
-
-    /// Validates, normalizes, and scrubs PII from a replay event.
-    fn process_replay_event(
-        &self,
-        payload: &Bytes,
-        config: &ProjectConfig,
-        client_ip: Option<NetIPAddr>,
-        user_agent: &RawUserAgentInfo<&str>,
-    ) -> Result<Annotated<Replay>, ReplayError> {
-        let mut replay =
-            Annotated::<Replay>::from_json_bytes(payload).map_err(ReplayError::CouldNotParse)?;
-
-        if let Some(replay_value) = replay.value_mut() {
-            replay::validate(replay_value)?;
-            replay::normalize(replay_value, client_ip, user_agent);
-        } else {
-            return Err(ReplayError::NoContent);
-        }
-
-        if let Some(ref config) = config.pii_config {
-            let mut processor = PiiProcessor::new(config.compiled());
-            processor::process_value(&mut replay, &mut processor, ProcessingState::root())
-                .map_err(|e| ReplayError::CouldNotScrub(e.to_string()))?;
-        }
-
-        let pii_config = config
-            .datascrubbing_settings
-            .pii_config()
-            .map_err(|e| ReplayError::CouldNotScrub(e.to_string()))?;
-        if let Some(config) = pii_config {
-            let mut processor = PiiProcessor::new(config.compiled());
-            processor::process_value(&mut replay, &mut processor, ProcessingState::root())
-                .map_err(|e| ReplayError::CouldNotScrub(e.to_string()))?;
-        }
-
-        Ok(replay)
     }
 
     /// Creates and initializes the processing state.
@@ -2286,7 +2153,7 @@ impl EnvelopeProcessorService {
             self.inner.config.clone(),
             self.inner.outcome_aggregator.clone(),
         );
-        self.process_replays(state)?;
+        replay::process(state, self.inner.config.clone())?;
         self.filter_profiles(state);
 
         if state.creates_event() {
@@ -2762,6 +2629,7 @@ mod tests {
     use insta::assert_debug_snapshot;
     use relay_base_schema::metrics::{DurationUnit, MetricUnit};
     use relay_common::glob2::LazyGlob;
+    use relay_dynamic_config::ProjectConfig;
     use relay_event_normalization::{
         MeasurementsConfig, NormalizeProcessor, RedactionRule, TransactionNameRule,
     };
